Implement simulateStaticDelegatecall via interface (#44)
As per @rmeissner's suggestion in #38. This is a much more elegant way of achieving the same thing (no code duplication). Unfortunately this is an API breaking change and we already published the npm package vor v3.0.0. Maybe we can deprecate this package and republish this without a major version update?

### Test Plan

unit test still passing.