Refactor storage accessible into smaller files. (#56)
This PR refactors `StorageAccessible` into smaller files, so that parts partial functionality can be added to a code-constrained contract (as discussed in the SC bi-weekly).

Specifically the files are:
- `StorageReadable`: abstract base contract containing `getStorageAt` implementation.
- `StorageSimulation`: containing the always reverting delegate call simulation method `simulateAndRevert`
- `ViewStorageAccessible`: `view` interface around the `StorageAccessible` contract

The `StorageAccessible` contract, through inheritance, has feature parity with the previous implementation.

The Solidity code we just moved around and is unchanged, with a single small, sneaky exception. We have empty statements for `StorageAccessible.simulate` parameters in order to avoid compiler warnings:
```
Warning: Unused function parameter. Remove or comment out the variable name to silence this warning.
  --> contracts/storage/StorageAccessible.sol:23:9:
   |
23 |         address targetContract,
   |         ^^^^^^^^^^^^^^^^^^^^^^


Warning: Unused function parameter. Remove or comment out the variable name to silence this warning.
  --> contracts/storage/StorageAccessible.sol:24:9:
   |
24 |         bytes calldata calldataPayload
   | 
```

Note that the above change has no effect on the resulting bytecode.

### Test Plan

The previous unit tests were moved into their individual files.